+++ core/model/src/main/java/com/wafflestudio/snutt2/core/model/data/PlaceTime.kt
@@ -21,18 +21,7 @@ enum class Day {
     SUNDAY
 }
 
-data class Time (
-    val timeInMinutes: Int
-) {
-    companion object {
-        fun fromHour(timeInHours: Int): Time = Time(
-            timeInHours * 60
-        )
-    }
 
-    val minute: Int get() = timeInMinutes % 60
-    val hour: Int get() = timeInMinutes / 60
-}
 
 data class Place(
     val name: String,
